Give Session its own module, make hvac use it
